selenium refactor cherry picked from container-frontend
