+++ selenium/Gemfile
@@ -12,9 +12,10 @@ source "http://rubygems.org"
 gem "json", "1.8.0"
 gem 'multi_json', '1.10.0'
 gem 'rubyzip', '1.0.0'
+gem "sinatra"
 
 group "test" do
-	gem "selenium-webdriver", "2.45.0"
+	gem "selenium-webdriver", ">= 2.45.0"
 	gem "rspec", "~> 3.3.0"
   gem "rspec-retry"
   gem "parallel_tests", "~> 1.3.9"
@@ -23,4 +24,12 @@ group "test" do
   gem "atomic", '= 1.0.1'
   gem "net-http-persistent", "2.8"
   gem "multipart-post", "1.2.0"
+
+end
+
+
+gemfile_local = File.join(File.dirname(__FILE__), 'Gemfile.local')
+if File.readable?(gemfile_local)
+  puts "Loading #{gemfile_local}..." if $DEBUG
+  instance_eval(File.read(gemfile_local))
 end
