+++ selenium/common.rb
@@ -9,486 +9,19 @@ require 'test_utils'
 require 'config/config-distribution'
 require 'securerandom'
 
+require_relative 'common/webdriver'
+require_relative 'common/helper_mixin'
+
 $server_pids = []
 $sleep_time = 0.0
 
 
-class RSpec::Core::Example
-  def passed?
-    @exception.nil?
-  end
-
-  def failed?
-    !passed?
-  end
-end
-
-
 module Selenium
-  module WebDriver
-    module Firefox
-      class Binary
-
-        # Searching the registry causes a EXCEPTION_ACCESS_VIOLATION under
-        # Windows 7.  Skip this step and just look for Firefox in the usual
-        # places.
-        def self.windows_registry_path
-          nil
-        end
-      end
-    end
-  end
-
   module Config
     def self.retries
       100
     end
   end
-
-end
-
-
-class Selenium::WebDriver::Driver
-
-  def wait_for_ajax
-    try = 0
-    while (self.execute_script("return document.readyState") != "complete" or
-      not self.execute_script("return window.$ == undefined || $.active == 0"))
-      if (try > Selenium::Config.retries)
-        puts "Retry limit hit on wait_for_ajax.  Going ahead anyway."
-        break
-      end
-
-      sleep(0.5)
-      try += 1
-    end
-
-    raise_javascript_errors
-  end
-
-  def find_paginated_element(*selectors)
-
-    start_page = self.current_url
-
-    try = 0
-    while true
-
-      begin
-        elt = self.find_element_orig(*selectors)
-
-        if not elt.displayed?
-          raise Selenium::WebDriver::Error::NoSuchElementError.new("Not visible (yet?)")
-        end
-
-        return elt
-
-      rescue Selenium::WebDriver::Error::NoSuchElementError
-        puts "#{test_group_prefix}find_element failed: trying to turn the page"
-        self.find_element_orig(:css => "a[title='Next']").click
-        retry
-      rescue Selenium::WebDriver::Error::NoSuchElementError
-        if try < Selenium::Config.retries
-          try += 1
-          sleep 0.5
-          self.navigate.to(start_page)
-          puts "#{test_group_prefix}find_paginated_element: #{try} misses on selector '#{selectors}'.  Retrying..." if (try % 5) == 0
-        else
-          raise Selenium::WebDriver::Error::NoSuchElementError.new(selectors.inspect)
-        end
-      end
-    end
-  end
-
-
-  def attempt(times, &block)
-
-    tries = times
-
-    begin
-      block.call(self)
-    rescue Exception => e
-      if tries > 0
-        tries -= 1
-        $sleep_time += 0.1
-        sleep 0.5
-        puts "Attempts remaining: #{tries}"
-        retry
-      else
-        raise e
-      end
-    end
-  end
-
-
-  alias :find_element_orig :find_element
-  def find_element(*selectors)
-    wait_for_ajax
-
-    try = 0
-    while true
-      begin
-        elt = find_element_orig(*selectors)
-
-        if not elt.displayed?
-          raise Selenium::WebDriver::Error::NoSuchElementError.new("Not visible (yet?)")
-        end
-
-        return elt
-      rescue Selenium::WebDriver::Error::NoSuchElementError, Selenium::WebDriver::Error::StaleElementReferenceError => e
-        if try < Selenium::Config.retries
-          try += 1
-          $sleep_time += 0.1
-          sleep 0.5
-          puts "#{test_group_prefix}find_element: #{try} misses on selector '#{selectors}'.  Retrying..." if (try % 5) == 0
-
-        else
-          puts "Failed to find #{selectors}"
-
-          if ENV['SCREENSHOT_ON_ERROR']
-            SeleniumTest.save_screenshot
-          end
-
-          raise e
-        end
-      end
-    end
-  end
-
-
-  def find_last_element(*selectors)
-    result = blocking_find_elements(*selectors)
-
-    result[result.length - 1]
-  end
-
-
-  def blocking_find_elements(*selectors)
-    # Hit with find_element first to invoke our usual retry logic
-    find_element(*selectors)
-
-    find_elements(*selectors)
-  end
-
-
-  def ensure_no_such_text(xpath, pattern, noError = false, noRetry = true )
-    wait_for_ajax
-    begin
-      element = self.find_element(:tag_name => "body").find_element_with_text(xpath, pattern, noError, noRetry)
-
-      if element.nil? or !element.displayed?
-        true
-      else
-        raise "Element was supposed to be absent: #{xpath} #{pattern}"
-      end
-    rescue Selenium::WebDriver::Error::NoSuchElementError => e
-      return true
-    end
-  end
-
-
-  def ensure_no_such_element(*selectors)
-    wait_for_ajax
-
-    begin
-      element = find_element_orig(*selectors)
-
-      if element.displayed?
-        raise "Element was supposed to be absent: #{selectors}"
-      else
-        true
-      end
-    rescue Selenium::WebDriver::Error::NoSuchElementError => e
-      return true
-    end
-  end
-
-
-  def click_and_wait_until_gone(*selector)
-    element = self.find_element(*selector)
-    element.click
-
-    begin
-      try = 0
-      while self.find_element_orig(*selector).equal? element
-        if try < Selenium::Config.retries
-          try += 1
-          $sleep_time += 0.1
-          sleep 0.5
-          puts "click_and_wait_until_gone: #{try} hits selector '#{selector}'.  Retrying..." if (try % 5) == 0
-        else
-          raise Selenium::WebDriver::Error::NoSuchElementError.new(selector.inspect)
-        end
-      end
-    rescue Selenium::WebDriver::Error::NoSuchElementError
-      nil
-    end
-  end
-
-  def generate_4part_id
-    Digest::MD5.hexdigest("#{Time.now}#{SecureRandom.uuid}#{$$}").scan(/.{6}/)[0...1]
-  end
-
-  def complete_4part_id(pattern, accession_id = nil)
-    # Was 4, but now that the input boxes are disabled this wasn't filling out the last 3.
-    accession_id ||= generate_4part_id
-    accession_id.each_with_index do |elt, i|
-      self.clear_and_send_keys([:id, sprintf(pattern, i)], elt)
-    end
-
-    accession_id
-  end
-
-
-  def find_element_with_text(xpath, pattern, noError = false, noRetry = false)
-    tries = 0
-
-    begin
-      self.find_element(:tag_name => "body").find_element_with_text(xpath, pattern, noError, noRetry)
-    rescue Selenium::WebDriver::Error::StaleElementReferenceError
-      if tries < Selenium::Config.retries
-        tries += 1
-        retry
-      end
-    end
-  end
-
-
-  def clear_and_send_keys(selector, keys)
-    Selenium::Config.retries.times do
-      begin
-        elt = self.find_element_orig(*selector)
-        elt.clear
-        elt.send_keys(keys)
-        break
-      rescue
-        $sleep_time += 0.1
-        sleep 0.3
-      end
-    end
-  end
-
-  def raise_javascript_errors
-    errors = $driver.execute_script("return window.hasOwnProperty('TEST_ERRORS') ? TEST_ERRORS : []")
-    raise "Javascript errors present: #{errors.inspect}" if errors.length > 0
-  end
-
-  def test_group_prefix
-    if ENV['TEST_ENV_NUMBER']
-      number = ENV['TEST_ENV_NUMBER'].empty? ? 1 : ENV['TEST_ENV_NUMBER']
-      "[#{number}] "
-    else
-      ""
-    end
-  end
-
-end
-
-
-class Selenium::WebDriver::Element
-
-  def select_option(value)
-    self.find_elements(:tag_name => "option").each do |option|
-      if option.attribute("value") === value
-        option.click
-        return
-      end
-    end
-
-    raise "Couldn't select value: #{value}"
-  end
-
-
-  def select_option_with_text(value)
-    self.find_elements(:tag_name => "option").each do |option|
-      if option.text === value
-        option.click
-        return
-      end
-    end
-
-    raise "Couldn't select value: #{value}"
-  end
-
-
-  def get_select_value
-    self.find_elements(:tag_name => "option").each do |option|
-      return option.attribute("value") if option.attribute("checked")
-    end
-
-    return ""
-  end
-
-  def nearest_ancestor(xpath)
-    self.find_element(:xpath => './ancestor-or-self::' + xpath + '[1]')
-  end
-
-
-  def containing_subform
-    nearest_ancestor('div[contains(@class, "subrecord-form-fields")]')
-  end
-
-
-  def find_last_element(*selectors)
-    result = find_elements(*selectors)
-
-    result[result.length - 1]
-  end
-
-
-  def find_element_with_text(xpath, pattern, noError = false, noRetry = false)
-    Selenium::Config.retries.times do |try|
-
-      matches = self.find_elements(:xpath => xpath)
-
-      begin
-        matches.each do | match |
-          return match if match.text =~ pattern
-        end
-      rescue
-        # Ignore exceptions and retry
-      end
-
-      if noRetry
-        return nil
-      end
-
-      $sleep_time += 0.1
-      sleep 0.5
-      puts "find_element_with_text: #{try} misses on selector ':xpath => #{xpath}'.  Retrying..." if (try % 10) == 0
-    end
-
-    return nil if noError
-
-    if ENV['SCREENSHOT_ON_ERROR']
-      SeleniumTest.save_screenshot
-    end
-
-    raise Selenium::WebDriver::Error::NoSuchElementError.new("Could not find element for xpath: #{xpath} pattern: #{pattern}")
-  end
-
-end
-
-
-
-def login(user, pass)
-  $driver.navigate.to $frontend
-  $driver.wait_for_ajax
-  $driver.find_element(:link, "Sign In").click
-  $driver.clear_and_send_keys([:id, 'user_username'], user)
-  $driver.clear_and_send_keys([:id, 'user_password'], pass)
-  $driver.find_element(:id, 'login').click
-  $driver.wait_for_ajax
-end
-
-
-def logout
-  $driver.navigate.to $frontend
-  ## Complete the logout process
-  user_menu = $driver.find_elements(:css, '.user-container .dropdown-menu.pull-right').first
-  if !user_menu || !user_menu.displayed?
-    $driver.find_element(:css, 'body').find_element(:css, '.user-container .btn.dropdown-toggle').click
-  end
-
-  $driver.find_element(:link, "Logout").click
-  $driver.find_element(:link, "Sign In")
-end
-
-
-# Drop a call to this into a Selenium test block to work with the browser
-# interactively.  Useful during test development!
-def debug_repl
-  $stderr.puts("Debug REPL active.  Type :eof to abort")
-  while true
-    line = $stdin.readline
-    break if line.strip == ":eof"
-    begin
-      puts eval(line).inspect
-    rescue
-      puts $!
-    end
-  end
-end
-
-
-module RepositoryHelperMethods
-
-  def create_test_repo(code, name, wait = true)
-    create_repo = URI("#{$backend}/repositories")
-
-    req = Net::HTTP::Post.new(create_repo.path)
-    req.body = "{\"repo_code\": \"#{code}\", \"name\": \"#{name}\"}"
-
-    response = admin_backend_request(req)
-    repo_uri = JSON.parse(response.body)['uri']
-
-
-    # Give the notification time to fire
-    sleep 5 if wait
-
-    @test_repositories ||= {}
-    @test_repositories[code] = repo_uri
-
-    [code, repo_uri]
-  end
-
-
-  def select_repo(code)
-    code = code.respond_to?(:repo_code) ? code.repo_code : code
-
-    $driver.find_element(:link, 'Select Repository').click
-    $driver.find_element(:css, '.select-a-repository').find_element(:id => "id").select_option_with_text(code)
-    $driver.find_element(:css, '.select-a-repository .btn-primary').click
-    if block_given?
-      $test_repo_old = $test_repo
-      $test_repo_uri_old = $test_repo_uri
-
-      $test_repo = code
-      $test_repo_uri = @test_repositories[code]
-      yield
-
-      $test_repo = $test_repo_old
-      $test_repo_uri = $test_repo_uri_old
-    end
-  end
-
-
-  def login_to_repo(user, pass, repo)
-    begin
-      logout
-    rescue # maybe we were already logged out
-    end
-
-    $driver.attempt(5) {|attempt|
-      login(user, pass)
-      select_repo(repo)
-    }
-  end
-
-
-  def generate_4part_id
-    Digest::MD5.hexdigest("#{Time.now}#{SecureRandom.uuid}#{$$}").scan(/.{6}/)[0...1]
-  end
-end
-
-
-
-class RepositoryHelper
-  def initialize
-    @test_repositories = {}
-  end
-
-  include RepositoryHelperMethods
-end
-
-
-$repository_helper = RepositoryHelper.new
-
-def create_test_repo(*args)
-  $repository_helper.create_test_repo(*args)
-end
-
-def select_repo(code, &block)
-  $repository_helper.select_repo(code, &block)
 end
 
 
@@ -507,7 +40,6 @@ def cleanup
     TestUtils::kill(pid)
   end
 end
-
 
 
 def selenium_init(backend_fn, frontend_fn)
